requirements.txt

asgiref==3.8.1
attrs==24.3.0
Automat==24.8.1
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.6
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
dataclasses==0.6
DateTime==5.5
defusedxml==0.7.1
Django==3.1.12
django-cors-headers==3.7.0
django-mongodb-engine==0.6.0
django-storages==1.10
djangorestframework==3.11.0
djangotoolbox==1.8.0
djongo==1.3.7
dnspython==2.6.1
eventlet==0.38.2
exceptiongroup==1.2.2
fake-useragent==2.0.3
Faker==33.1.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-SocketIO==5.5.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
greenlet==3.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
mysql-connector-python==9.1.0
netifaces==0.10.6
numpy==1.24.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3
parsel==1.9.1
pillow==10.2.0
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pymongo==3.11.4
pyOpenSSL==24.3.0
pyparsing==3.2.0
pypiwin32==223
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pyttsx3==2.91
pytz==2024.1
pywin32==306
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rfc3986==1.5.0
rsa==4.9
Scrapy==2.12.0
scrapy-fake-useragent==1.4.4
scrapy-rotating-proxies==0.6.2
selenium==4.27.1
service-identity==24.2.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
socketio==0.2.1
sortedcontainers==2.4.0
sqlparse==0.2.4
sympy==1.13.3
tldextract==5.1.3
tomli==2.2.1
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.11.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.2
w3lib==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
zope.interface==7.0.3
